Move backup images to separate folders, and Google Sync
To prevent image folders from becoming clogged, backup folders and "*.kra " files folders were created. The files will be deleted when creating the build, but let the players viewing the files know that one of the developers used Krita®)

Due to syncing with Google Drive, additional files were created with a note about sharing in each folder. Unpleasant, but not critical.